+++ example/radio/index.vue
@@ -5,6 +5,7 @@
       <title title="wxc-radio"></title>
       <category title="Radio可选"></category>
       <wxc-radio :list="list"
+                 :config="config"
                  @wxcRadioListChecked="wxcRadioListChecked"></wxc-radio>
       <text class="radio-text">checkedItem: {{checkedInfo}}</text>
       <category title="Radio不可选"></category>
@@ -55,6 +56,11 @@
         { title: '未选不可修改', value: 5, disabled: true },
         { title: '已选不可修改', value: 6, disabled: true, checked: true }
       ],
+      config:{
+        checkedIcon:'https://gw.alicdn.com/tfs/TB1TARHdf6H8KJjSspmXXb2WXXa-31-30.png',
+        disabledIcon:'https://gw.alicdn.com/tfs/TB1PtN3pwMPMeJjy1XdXXasrXXa-72-72.png',
+        checkedColor: '#000000'
+      },
       checkedInfo: { title: '选项2', value: 2, oldIndex: -1, index: 1 }
     }),
     created () {