+++ packages/wxc-checkbox/checkboxList.vue
@@ -5,6 +5,7 @@
   <div>
     <wxc-checkbox v-for="(item,i) in list"
                   v-bind="item"
+                  :config="config"
                   @wxcCheckBoxItemChecked="wxcCheckBoxItemChecked"
                   :key="i"></wxc-checkbox>
   </div>
@@ -19,6 +20,10 @@
       list: {
         type: Array,
         default: () => ([])
+      },
+      config: {
+        type: Object,
+        default: () => ({})
       }
     },
     data: () => ({
@@ -37,7 +42,7 @@
         if (e.checked) {
           this.checkedList.push(e.value);
         } else {
-          var index = this.checkedList.indexOf(e.value);
+          const index = this.checkedList.indexOf(e.value);
           this.checkedList.splice(index, 1);
         }
         this.$emit('wxcCheckBoxListChecked', { checkedList: this.checkedList })