+++ packages/wxc-result/type.js
@@ -5,25 +5,25 @@
 
 export default {
   errorPage: {
-    pic: 'https://gtms01.alicdn.com/tfs/TB1HH4TSpXXXXauXVXXXXXXXXXX-320-320.png',
-    content: '抱歉出错了，飞猪正在全力解决中',
+    pic: 'https://img.alicdn.com/tfs/TB17blphfDH8KJjy1XcXXcpdXXa-320-320.png',
+    content: '抱歉出错了，我们正在全力解决中',
     button: '再试一次',
     title: '出错啦'
   },
   noGoods: {
-    pic: 'https://gw.alicdn.com/tfs/TB1QXlEQXXXXXcNXFXXXXXXXXXX-320-320.png',
+    pic: 'https://img.alicdn.com/tfs/TB1mPWEeOqAXuNjy1XdXXaYcVXa-320-320.png',
     content: '主人，这里什么都没有找到',
     button: '再试一次',
     title: '暂无商品'
   },
   noNetwork: {
-    pic: 'https://gw.alicdn.com/tfs/TB1rs83QXXXXXcBXpXXXXXXXXXX-320-320.png',
+    pic: 'https://img.alicdn.com/tfs/TB1jkA5g9_I8KJjy0FoXXaFnVXa-320-320.png',
     content: '哎呀，没有网络了......',
     button: '刷新一下',
     title: '无网络'
   },
   errorLocation: {
-    pic: 'https://gw.alicdn.com/tfs/TB1rs83QXXXXXcBXpXXXXXXXXXX-320-320.png',
+    pic: 'https://img.alicdn.com/tfs/TB1zXXahhrI8KJjy0FpXXb5hVXa-320-320.png',
     content: '哎呀，定位失败了......',
     button: '刷新一下',
     title: '定位失败'